refactor client package: - testutils.go have the utils required for vochaintest - zk.go have the artifacts required for anon voting and the api - api.go continues to have the methods used for interacting   with the exposed AP
cmd/vochaintest changed accordingly

This commit improves the clarity and ease of use of the client pkg
